+++ integrations/aws-v3/aws/core/exporters/codedeploy/__init__.py
@@ -1,3 +1,8 @@
+from aws.core.exporters.codedeploy.application import CodeDeployApplicationExporter
+from aws.core.exporters.codedeploy.application.models import (
+    SingleCodeDeployApplicationRequest,
+    PaginatedCodeDeployApplicationRequest,
+)
 from aws.core.exporters.codedeploy.deployment_group.exporter import CodeDeployDeploymentGroupExporter
 from aws.core.exporters.codedeploy.deployment_group.models import (
     SingleCodeDeployDeploymentGroupRequest,
@@ -5,6 +10,9 @@ from aws.core.exporters.codedeploy.deployment_group.models import (
 )
 
 __all__ = [
+    "CodeDeployApplicationExporter",
+    "SingleCodeDeployApplicationRequest",
+    "PaginatedCodeDeployApplicationRequest",
     "CodeDeployDeploymentGroupExporter",
     "SingleCodeDeployDeploymentGroupRequest",
     "PaginatedCodeDeployDeploymentGroupRequest",