+++ integrations/aws-v3/main.py
@@ -43,6 +43,10 @@ from aws.core.exporters.elasticache import ElastiCacheClusterExporter
 from aws.core.exporters.elasticache.cluster.models import PaginatedCacheClusterRequest
 from aws.core.exporters.ec2.volume import EbsVolumeExporter
 from aws.core.exporters.ec2.volume.models import PaginatedEbsVolumeRequest
+from aws.core.exporters.codedeploy import CodeDeployApplicationExporter
+from aws.core.exporters.codedeploy.application.models import (
+    PaginatedCodeDeployApplicationRequest,
+)
 from aws.core.exporters.codedeploy import CodeDeployDeploymentGroupExporter
 from aws.core.exporters.codedeploy.deployment_group.models import PaginatedCodeDeployDeploymentGroupRequest
 from aws.core.helpers.utils import is_access_denied_exception
@@ -263,6 +267,18 @@ async def resync_elasticache_cluster(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
 async def resync_ec2_volume(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     service = ResyncAWSService(
         kind, EbsVolumeExporter, PaginatedEbsVolumeRequest, regional=True
+    )
+    async for batch in service:
+        yield batch
+
+
+@ocean.on_resync(ObjectKind.CODEDEPLOY_APPLICATION)
+async def resync_codedeploy_application(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
+    service = ResyncAWSService(
+        kind,
+        CodeDeployApplicationExporter,
+        PaginatedCodeDeployApplicationRequest,
+        regional=True,
     )
     async for batch in service:
         yield batch