+++ integrations/custom/CHANGELOG.md
@@ -8,6 +8,43 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- towncrier release notes start -->
 
+## 0.6.49 (2026-06-09)
+
+
+### Improvements
+
+- Bumped ocean version to ^0.43.16
+
+
+## 0.6.48 (2026-06-07)
+
+
+### Improvements
+
+- Bumped ocean version to ^0.43.15
+
+
+## 0.6.47 (2026-06-04)
+
+### Improvements
+
+- Add docs to custom SaaS installation types
+
+## 0.6.46 (2026-06-03)
+
+
+### Improvements
+
+- Bumped ocean version to ^0.43.14
+
+
+## 0.6.45 (2026-06-03)
+
+
+### Improvements
+
+- Move custom ocean to Build Your Own Integration
+
 ## 0.6.44 (2026-06-03)